feedback: place alert warning below section header
