+++ frontend/packages/ux-editor/src/components/Properties/Properties.module.css
@@ -13,7 +13,3 @@
   padding-block: var(--fds-spacing-5) 0;
   padding-inline: 0;
 }
-
-.alertWrapper {
-  margin: var(--fds-spacing-5);
-}
