+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/PropertiesHeader.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import classes from './PropertiesHeader.module.css';
 import { formItemConfigs } from '../../../data/formItemConfig';
 import { QuestionmarkDiamondIcon } from '@studio/icons';
-import { StudioSectionHeader } from '@studio/components';
-
+import { StudioAlert, StudioSectionHeader } from '@studio/components';
 import { getComponentHelperTextByComponentType } from '../../../utils/language';
 import { useTranslation } from 'react-i18next';
 import { EditComponentIdRow } from './EditComponentIdRow';
@@ -11,6 +10,7 @@ import type { FormItem } from '../../../types/FormItem';
 import { ComponentType } from 'app-shared/types/ComponentType';
 import { EditLayoutSetForSubform } from './EditLayoutSetForSubform';
 import { ComponentMainConfig } from './ComponentMainConfig';
+import { isComponentDeprecated } from '@altinn/ux-editor/utils/component';
 
 export type PropertiesHeaderProps = {
   formItem: FormItem;
@@ -41,6 +41,11 @@ export const PropertiesHeader = ({
           title: t('ux_editor.component_help_text_general_title'),
         }}
       />
+      {isComponentDeprecated(formItem.type) && (
+        <StudioAlert size='sm' className={classes.alertWrapper} severity='warning'>
+          {t(`ux_editor.component_properties.deprecated.${formItem.type}`)}
+        </StudioAlert>
+      )}
       <div className={classes.content}>
         <EditComponentIdRow component={formItem} handleComponentUpdate={handleComponentUpdate} />
         {formItem.type === ComponentType.Subform && (
